backend/requirements/requirements.txt
fastapi
uvicorn
pydantic
pandas
pyzabbix
google-generativeai
sqlalchemy
pymysql
python-dotenv